Beta of JAMES 1.2.
Major changes is the separation of mailet API.  Interfaces defined in org.apache.mailet, and Avalon/JAMES implementation handled by org.apache.james.core and other classes in various packages (as appropriate).

Longer term, the mailet API will hopefully be contained in a separate JAR and the source code for this jar moved to a separate part of the CVS tree.  For the moment, the build.xml compiles all code into one jar.


git-svn-id: https://svn.apache.org/repos/asf/james/server@107158 13f79535-47bb-0310-9956-ffa450edef68