Introduced the concept of a "sourceId" on MimeMessageSource.  This should be a unique ID that represents where this source is retrieving this message.  In theory equals() should be written to use this (if we needed that method).
This then allows the AvalonMailRepository to see where it is retrieving a message compared to where it is saving it.  If it sees a message where it is trying to save to the same place it would be retrieving from, it will not do so unless the message has changed.  If the message has changed, the message has in fact been loaded into memory and is also logical for the message to be saved.

This should allow the mail repository to handle saves reliably rather than worrying about calling store() on a mail object that hasn't changed.


git-svn-id: https://svn.apache.org/repos/asf/james/server@107873 13f79535-47bb-0310-9956-ffa450edef68