This is a correction to the SMTP AUTH code to close an open relay hole.
In short, there was an unnecessary clause in the SMTP AUTH code that
excluded emails with null senders from validation, under the premise
that these are delivery failure notifications.

This is incorrect.  Delivery failure notifications generated by the
James server are placed directly on the outgoing spool.  With SMTP
AUTH configured external delivery failure notifications should 
be routed through the James server if and only if they are 
intended for delivery to the James server.  Most critically,
before this patch a James server in a minimal SMTP AUTH config is
recognized as an open relay by ORDB.org.  After, it is not.


git-svn-id: https://svn.apache.org/repos/asf/james/server/trunk@107971 13f79535-47bb-0310-9956-ffa450edef68