Fixed a major bug that was exposed by a recent change to JDBCMailRepository to restore dbfile capability.
This bug caused configurations of one repository to be used by another differently configured
repository of the same class.  For some configurations, this means that a formerly db repository
behaved as if it were a dbfile repository.

Before using this code make sure that you don't need anything in var/dbmail.


git-svn-id: https://svn.apache.org/repos/asf/james/server/trunk@108115 13f79535-47bb-0310-9956-ffa450edef68