Many changes to the NNTP server.
Corrected a number of outstanding bugs, including issues with command parsing,
NNTP authentication/authorization, and some group handling.

Added substantial documentation.

A number of issues may remain - a number of TODOs have been added to draw attention
to some potential problems.

Note that deployment of this distribution will require a change to the assembly.xml
as per the change to james-assembly.xml


git-svn-id: https://svn.apache.org/repos/asf/james/server@108124 13f79535-47bb-0310-9956-ffa450edef68