1) The MailImpl instance sent to James now includes the remote address and remote host of the n'th relay in the delivery chain as reported by the 'from' field of the 'RECEIVED' headers. This is equivalent to the remote address and remote host set in the MailImpl by the SMTPHandler deduced from the socket of the connected server.
The n'th relay is specified by the new fetchmail configuration parameter
'remoteReceivedHeaderIndex'. This is explained in the configuration file and
updated xdoc.

The InSpammerBlackList matcher validates against the remote address in the
MailImpl. Prior to this patch it was having no effect and all mail was
considered valid.

2) Made logging at the INFO level less verbose and logging at the DEBUG
level more detailed.

Submitted by:	Steve Brewin


git-svn-id: https://svn.apache.org/repos/asf/james/server/branches/branch_2_1_fcs@108995 13f79535-47bb-0310-9956-ffa450edef68