Changed getSMTPHostAddresses to call findMXRecords, which already knew how to return a collection of  MX hosts or the host if there weren't any MX records defined.
Changed to use InetAddress.getAllByAddress rather than invoking DNS lookup directly.  The former is more robust, and handles CNAME RR (even if it is recommended that an MX never have a CNAME on the RHS).

Left MxSorter in place so that we can discuss fixing and keeping it, or removing it.  Comments indicate the change(s) to fix it.


git-svn-id: https://svn.apache.org/repos/asf/james/server/branches/branch_2_1_fcs@109088 13f79535-47bb-0310-9956-ffa450edef68