Fix JAMES-9.
JamesSpoolManager.dispose() now tells the spooler threads to terminate
and then waits (up to 1 minute) for them to complete.

Commented out the use of ThreadPool.  It was never actually used as a
thread pool.  All of the threads were created at initialization, and
remained dedicated for the duration of the application.  The way a
thread pool is normally used is for a series of short-lived runners.
Future revisions to the spooler could use a thread pool by refactoring
the run() method, and having a small number of threads that get mail
from the spool, and dispatch them to workers in a thread pool for
processing.


git-svn-id: https://svn.apache.org/repos/asf/james/server/branches/branch_2_1_fcs@109097 13f79535-47bb-0310-9956-ffa450edef68