First pass for JAMES-611:
Removed finalizers from MimeMessageCopyOnWriteProxy and MimeMessageInputStreamSource: the finalizers were simple calls to dispose(), nothing more.

Fixed one MailImpl constructor to handle exceptions during the construction and related resource disposal: we should review every other constructor (this one was leaking on MessageSizeException).

Added a check in MimeMessageInputStreamSource to remove the just created file if the input stream throw an IOError (was leaking on MessageSizeException)

Added a check in DeleCmdHandler to make sure we dispose the loaded mail when replacing it with the DELETED object (must be reviewed: maybe it would be better to dispose it when we close the connection and delete the messages from the repository).

POP3Handler: added a loop over backupUserMailbox when resetting the handler to make sure objects are disposed (not sure this is the right thing to do)

Updated a lot of tests to make sure they dispose messages/mails they create.


git-svn-id: https://svn.apache.org/repos/asf/james/server/trunk@441785 13f79535-47bb-0310-9956-ffa450edef68