First attempt at creating a delegating socket-server. - changed some visibility from AbstractJamesHandler to allow my extension. - Introduced ProtocolHandler and ProtocolHandlerHelper interfaces. - created a DelegatingJamesHandler that expose a ProtocolHandlerHelper service to the managed "ProtocolHandler" (received in the constructor). - created a DelegatingSMTPServer that is identical to SMTPServer but manages a pool of DelegatingJamesHandler (containing SMTPProtocolHandlers) instead  of SMTPHandlers. - created an SMTPProtocolHandler by cloning the current SMTPHandler.
The main advantage is that SMTPProtocolHandler has no knowledge of cornerstone/excalibur because they are incapsulated in the DelegatingJamesHandler .
The used interfaces are meant to make it really simple the porting of our current services to this structure. We can then refactor the interface to expose better names for methods/services.
We could also move the ProtocolHandler/ProtocolHandlerHelper to a lower layer.

This code is *untested*, ATM.

git-svn-id: https://svn.apache.org/repos/asf/james/server@685510 13f79535-47bb-0310-9956-ffa450edef68